Publicly expose service context's trace ID (#311)
## Motivation

Users may want to access the task local `ServiceContext`'s trace ID to,
for example, include it in response headers as part of a middleware.
This is not currently possible using `1.0.0-alpha.1` because the entire
span context stored inside `ServiceContext` is internal.

## Modifications

We explored various ways of exposing the trace ID in #293 and came to an
agreement that we want to expose as little as possible and ensure that
our dependency on
[swift-otel/swift-w3c-trace-context](https://github.com/swift-otel/swift-w3c-trace-context)
remains an implementation detail. Therefore, we only expose the trace
ID's String representation.

## Result

Users can now obtain a String representation of a given
`ServideContext`'s trace ID.

Closes #293