Use consistent withTimeout logic across processors (#312)
## Motivation

We have a `withTimeout` helper which was useful, but needed some
improvements: it conflated task cancellation with timeout. Additionally,
this was not being consistently used across the project, with the
processors having their own task group for the timeouts. This led to
much more complicated run methods than necessary. Finally, the
BatchSpanProcessor and the BatchLogRecordProcessor have essentially the
same semantics, but their internal workings were very different.

## Modifications

- Update withTimeout helper to cover corner cases
- Adopt withTimeout in BatchLogRecordProcessor
- Wait for buffer in to fill in BatchLogRecordProcessorTests
- Adopt withTimer in BatchSpanProcessor
- Align BatchLogRecordProcessor with BatchSpanProcessor

## Result

- Better tests
- Simpler processor logic
- Consisten processor logic