Consume response body in HTTP retries test for clean client shutdown (#322)
## Motivation

The tests for HTTP retries were causing flakey errors on teardown
because the AHC wasn't shutting down cleanly. We attempted to address
this in #319 but it turns out it was actually because we were not
consuming the response body, which led to a race when the client was
being torn down.

## Modifications

Consume response body in HTTP retries test for clean client shutdown

## Result

Test is now consistently passing on Linux too.

## Notes

Tested this in a loop on Linux in a container. And credit to @fabianfett
to helping me resolve this one.