Change span processor requirements to use non-async onStart and onEnd (#317)
## Motivation

The `protocol OTelSpanProcessor` defines requirements for `onStart` and
`onEnd` that are async. In the context of the `BatchSpanProcessor`, this
is fulfilled by an actor method. However, `onStart` and `onEnd` are
called on the hot path so we'd like these to be immediate, deferring any
actual processing and exporting work to the background. Right now,
they'll wait for the actor to be free, which might currently be doing an
export, for example.

The `protocol OTelLogRecordProcessor` does not have this drawback since
its `onEmit` requirement is non-async and so the
`BatchLogRecordProcessor` fulfils this with a `nonisolated` method which
yields into an `AsyncStream` which is consumed by a child task in the
processor's run method. This is much better, since we don't want calls
to `Logger.log` to block the hot path.

The `BatchSpanProcessor` and `BatchLogRecordProcessor` share a similar
goal and there's no real reason for this difference.

Before we made these protocols and types internal, we would have had API
considerations, but now we can make this change.

## Modifications

- Change span processor requirements to use non-async onStart and onEnd

## Result

Batch span processor does not block hot path when span starts or ends.