Drop spans/logs when batch processor buffer is full and emit warning (#324)
## Motivation

Swift OTel has never dropped spans when the queue size was reached,
rather it just triggered an additional export. However the spec says
that the batch span and log record processors should drop when the queue
is full[^1][^2].

## Modifications

- Drop spans/logs when batch processor buffer is full and emit warning

## Result

Batch span and log record processors drop spans and logs when the queue
is full.

[^1]:
https://opentelemetry.io/docs/specs/otel/trace/sdk/#batching-processor
[^2]:
https://opentelemetry.io/docs/specs/otel/logs/sdk/#batching-processor