Use an internal ByteBufferWrapper type to avoid a copy through Data (#330)
## Motivation

In the OTLP/HTTP exporters, we currently have a copy to deserialize the
response type because NIO ByteBuffer does not conform to
SwiftProtobufContiguousBytes. We cannot add a retroactive conformance
without making the import of NIO public. But we can solve this with an
internal wrapper type that conforms.

## Modifications

Use an internal ByteBufferWrapper type to avoid a copy through Data

## Result

Increased performance of OTLP/HTTP exporter by removing an extra copy of
response body.