Also use conditional FoundationEssentials imports for typed imports (#334)
## Motivation

On platforms that split Foundation into FoundationEssentials, we want to
only require our users link the latter. We tried to do this in #314 but
this missed "scoped imports", e.g. `import struct Foundation.Data`.

## Modifications

Also use conditional FoundationEssentials imports for typed imports

## Result

Remove requirement for full Foundation.

## Notes

I'll see if I can work on some CI for this too, in a followup PR.