Add CI for Foundation imports without FoundationEssentials equivalent (#336)
## Motivation

Even though some of our dependencies are not yet
`FoundationEssentials`-clean, we don't want to regress the fact that
this repo does not force the full Foundation dependency on downstream
packages itself.

## Modifications

Add CI for Foundation imports without FoundationEssentials equivalent.

## Result

CI will catch if we have a Foundation import without an equivalent
FoundationEssentials import.

## Notes

Here's an example of this failing if I add an `import struct
Foundation.Data` locally:

```console
% scripts/check-foundation-essentials-dependency.sh
** ERROR: Sources/OTel/OTLPHTTP/OTLPHTTPExporter.swift: import has no FoundationEssentials equivalent: import struct Foundation.Data
** ERROR: ❌ Found 1 errors.
```